+++ Code/Mantid/Vates/VatesAPI/src/RebinningCutterPresenter.cpp
@@ -245,8 +245,8 @@ std::vector<boost::shared_ptr<Mantid::Geometry::IMDDimension> > getDimensions(
   std::vector<boost::shared_ptr<IMDDimension> > dimensionVec;
 
   //Extract dimensions
-  int nDimensions = dimensionsXML->length();
-  for (int i = 0; i < nDimensions; i++)
+  size_t nDimensions = dimensionsXML->length();
+  for (size_t i = 0; i < nDimensions; i++)
   {
     Poco::XML::Element* dimensionXML = static_cast<Poco::XML::Element*> (dimensionsXML->item(i));
     DimensionFactory factory(dimensionXML);
